ref(cli): move the recipe glob and renderer into src/prompts
`commands/help.ts` owned the `import.meta.glob` over `help/*.txt`, the
canonical/anonymous map build, the `TOPIC_INPUT_SCHEMAS` table, and
`renderRecipe`. That put the recipe source and its interpolation behind a
citty command, so nothing else could reach the text without duplicating it.

Move all of it to `src/prompts/recipes.ts` and have `help` and `onboard`
call `getRecipe`. The module carries no CLI runtime: embedded text,
`sprintf-js`, `applyCliInvocation`, and the two leaf Zod input schemas.

Pure refactor. `help` output is byte-identical, verified across 38
captures covering all 18 canonical topics, their `--anonymous` variants,
the topic index, and the unknown-topic error path, with a zero-byte diff
before and after.

Unit 1 of 2 for export-knowledge-prompts.
